rename index-kappacore
