+++ package.json
@@ -21,7 +21,7 @@
     "kappa-core": "^3.0.2",
     "kappa-private": "^1.0.0",
     "kappa-view-kv": "^2.0.1",
-    "kappa-view-pull-query": "^1.0.0",
+    "kappa-view-pull-query": "^1.1.0",
     "level": "^5.0.1",
     "map-filter-reduce": "^3.2.2",
     "memdb": "^1.3.1",
